Tic-Tac-Toe game
implemented all the functionalities of the classic game of Tic-Tac-Toe:
first line / column / diagonal with three of the same elements will win, if there is no such condition, the game concludes with a tie.
Players will click a button at a time, X starts first.
At the end of the game, a button will appear that allows the users to play again, without the need to restart the page.